requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP Client
httpx==0.25.1
requests==2.31.0

# Data Parsing and Validation
pydantic==2.5.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Caching
redis==5.0.1
cachetools==5.3.2

# Environment and Configuration
python-dotenv==1.0.0
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

# Utilities
python-json-logger==2.0.7
tenacity==8.2.3

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0